Update mnist_test.py
Thankyou for this code, it is helpful to a beginner with Caffe. Caffe only shows how to train the model but not how to use it. Typical for ML/AI people, they seem to think everything is obvious.

You don't use Image or matplotlib.pyplot so you don't need to import them.

Also, which "Image" is that?